backend/requirements.txt
django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.2
python-dotenv==1.0.1
requests==2.32.3
dnspython==2.6.1
pillow==10.4.0
pypdf==4.2.0
whitenoise==6.7.0
gunicorn==22.0.0
psycopg2-binary==2.9.9
